backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
requests==2.32.3
python-dotenv==1.0.1
